Bug #37044: Read overflow in opt_range.cc found during "make test"
The code was using a special global buffer for the value of IS NULL ranges.
This was not always long enough to be copied by a regular memcpy. As a 
result read buffer overflows may occur.
Fixed by setting the null byte to 1 and setting the rest of the field disk image
to NULL with a bzero (instead of relying on the buffer and memcpy()).
--BZR--
revision-id: joro@sun.com-20090824122803-1d5jlaysjc7a7j6q
property-branch-nick: B37044-5.0-bugteam
property-rebase-of: joro@sun.com-20090824122803-h2dxfi1x5lhqycti
testament3-sha1: 1d611d0467c8775d8e3863937519803db6b91e3c