branches/zip: recv_recover_page_func(): Write the log sequence number to the compressed page, if there is one.  Previously, the function only wrote the LSN to the uncompressed page.
It is not clear why recv_recover_page_func() is updating FIL_PAGE_LSN
in the buffer pool.  The log sequence number will be stamped on the
page when it is flushed to disk, in buf_flush_init_for_writing().
I noticed this inconsistency when analyzing Issue #313, but this patch
does not fix it.  That is no surprise, since FIL_PAGE_LSN should only
matter on disk files, not in the buffer pool.
--BZR--
revision-id: svn-v4:16c675df-0fcb-4bc9-8058-dcc011a37293:branches/zip:5752
testament3-sha1: a0d40f641dc98def2badc8ef3a768b9d24f6cd3a