Bug #47106: Crash / segfault on adding EXPLAIN to a non-crashing  query
The fix for bug 46749 removed the check for OUTER_REF_TABLE_BIT 
and substituted it for a check on the presence of 
Item_ident::depended_from.
Removing it altogether was wrong : OUTER_REF_TABLE_BIT should 
still be checked in addition to depended_from (because it's not 
set in all cases and doesn't contradict to the check of depended_from).
Fixed by returning the old condition back as a compliment to the 
new one.
--BZR--
revision-id: joro@sun.com-20090918093408-154nmd4mw2aprtq4
property-branch-nick: B47106-5.0-bugteam
property-rebase-of: joro@sun.com-20090918093408-0puszc4djdw7m98a
testament3-sha1: 277bcbf1f0dcd3428bc2f98c6a19839fdf8ec3ca