BUG#47312: RBR: Disabling key on slave breaks replication: HA_ERR_WRONG_INDEX
In RBR, disabling keys on slave table will break replication when
updating or deleting a record. When the slave thread tries to
find the row, by searching in the storage engine, it checks
whether the table has a key or not. If it has one, then the slave
thread uses it to search the record.
      
Nonetheless, the slave only checks whether the key exists or not,
it does not verify if it is active. Should the key be
disabled (eg, DBA has issued an ALTER TABLE ... DISABLE KEYS)
then it will result in error: HA_ERR_WRONG_INDEX.
      
This patch addresses this issue by making the slave thread also
check whether the key is active or not before actually using it.
--BZR--
revision-id: luis.soares@sun.com-20090927210247-mc0tbx8uxsuojkch
property-branch-nick: mysql-5.1-bugteam
testament3-sha1: 7eb29e5d44ee4c4e8596b5e1284b2342567261dc