BUG#42928: binlog-format setting prevents server from start if binary logging is disabled
NOTE: this is the backport to next-mr.
            
If one sets binlog-format but does NOT enable binary log, server
refuses to start up. The following messages appears in the error log:
            
090217 12:47:14 [ERROR] You need to use --log-bin to make
--binlog-format work.  
090217 12:47:14 [ERROR] Aborting
            
This patch addresses this by making the server not to bail out if the
binlog-format is set without the log-bin option. Additionally, the
specified binlog-format is stored, in the global system variable
"binlog_format", and a warning is printed instead of an error.
--BZR--
revision-id: luis.soares@sun.com-20090929141243-4u503k8p3tvpr0a7
property-branch-nick: mysql-5.1-rep+2-to-push
testament3-sha1: 66ea5bf620efeb8c5ca4d41a1512e864a77cd750