Bug#46944 Internal prepared XA transction XIDs are not           removed if server_id changes
When MySQL crashes (or a snapshot is taken which simulates
a crash), then it is possible that internal XA
transactions (used to sync the binary log and InnoDB)
can be left in a PREPARED state, whereas they should be
rolled back.  This is done when the server_id changes
before the restart occurs.  

This patch releases he restriction that the server_id
should be consistent if the XID is to be considerred
valid. The rollback phase should then be able to
clean up all pending XA transactions.
--BZR--
revision-id: kristofer.pettersson@sun.com-20091012124600-edi4kbyks7cnjnsx
property-branch-nick: 51-bug46944
testament3-sha1: c0b5f6bebb983473e7707e9cd3b8d4176a938dcd