Bug #38124 "general_log_file" variable silently unset when using expression
When assigning the new string value to the variable, the
Item::str_value member was used.  This is not according to
the protocol.  str_value is an internal member used for
temporary assignments, and is not consistently set for all
string operations.  It is set for constant strings, so it would
work in these cases, but not for string functions (concat,
substr, etc.)
                  
The correct approach is to use Item::val_str(..) to evaluate
and retrieve the string.

Backport from 6.0-codebase

6.0-codebase revno: 2617.31.17

--BZR--
revision-id: magne.mahre@sun.com-20091015090931-z89vxyj9n63yxakp
property-branch-nick: mysql-next-mr-runtime
testament3-sha1: c4a99ac073d951cd179cd3e18c1c4d35ad47d930