Bug#47627 SET @@{global.session}.local_variable in stored routine causes crash
Adding @@session and @@global prefixes to a
declared variable in a stored procedure the server
would lead to a crash.

The reason was that during the parsing of the
syntactic rule 'option_value' an uninitialized
set_var object was pushed to the parameter stack
of the SET statement. The parent rule
'option_type_value'  interpreted the existence of
variables on the parameter stack as an assignment
and wrapped it in a sp_instr_set object.

As the procedure later was executed an attempt
was made to run the method 'check()' on an
uninitialized member object (NULL value) belonging
to the previously created but uninitialized object.

--BZR--
revision-id: kristofer.pettersson@sun.com-20091019074333-g2ces0lo4c2ejar7
property-branch-nick: 51-bug47627
property-file-info: ld7:file_id64:sp1f-sql_yacc.yy-19700101030959-wvn4qyy2drpmge7kaq3dysprbhlrv27j7:message301:* Assign the option_type at once since it is needed by the next
property-file-info:   parsing rule (internal_variable_name)
property-file-info: * Rearranged the if statement to reduce negations and gain more
property-file-info:   clarity of code.
property-file-info: * Added check for option_type to better detect if current
property-file-info:   variable is a SP local variable or a system variable.
property-file-info: 4:path15:sql/sql_yacc.yyee
testament3-sha1: 833638c5afb3c2ce71666650deabe905126027be