BUG#48216 Replication fails on all slaves after upgrade to 5.0.86 on master
When a sessione is closed, all temporary tables of the session are automatically 
dropped and are binlogged. But it will be binlogged with wrong database names when
the length of the temporary tables' database names are greater than the 
length of the current database name or the current database is not set.

Query_log_event's db_len is forgot to set when Query_log_event's db is set.
This patch wrote code to set db_len immediately after db has set.



--BZR--
revision-id: li-bing.song@sun.com-20091103090041-zj7nedx6ok5jgges
property-branch-nick: mysql-5.0-bugteam
testament3-sha1: 3d638fafc1fef78e38729fc079c73a3121394708