Applying InnoDB plugin snashot
Detailed revision comments:

r6048 | vasil | 2009-10-09 08:42:55 +0300 (Fri, 09 Oct 2009) | 16 lines
branches/zip:

When scanning a directory readdir() is called and stat() after it,
if a file is deleted between the two calls stat will fail and the
whole precedure will fail. Change this behavior to continue with the
next entry if stat() fails because of nonexistent file. This is
transparent change as it will make it look as if the file was deleted
before the readdir() call.

This change is needed in order to fix
https://svn.innodb.com/mantis/view.php?id=174
in which we need to abort if os_file_readdir_next_file()
encounters "real" errors.

Approved by:	Marko, Pekka (rb://177)

--BZR--
revision-id: svoj@sun.com-20091103095931-20wa5wnr23fwipak
property-branch-nick: mysql-5.1-bugteam
testament3-sha1: 22cf833dc3977a08ff33099905fe2a9a48857f40