WL#2687 WL#5072 BUG#40278 BUG#47175
Non-transactional updates that take place inside a transaction present problems
for logging because they are visible to other clients before the transaction
is committed, and they are not rolled back even if the transaction is rolled
back. It is not always possible to log correctly in statement format when both
transactional and non-transactional tables are used in the same transaction.

In the current patch, we ensure that such scenario is completely safe under the
ROW and MIXED modes.
--BZR--
revision-id: alfranio.correia@sun.com-20091103190256-637o8qxlveikrt3i
property-branch-nick: mysql-5.1-rep+3
testament3-sha1: 7f6a8d72edae00f7b19adb6fd15bb34013f23a3f