Bug#48031: mysql_secure_installation -- bash bug regarding passwords with   special chars
This script failed when the user tried passwords with multiple spaces, \, # or
' characters.  Now proper escaping and quoting is used in all contexts.

This problem occurs in the Perl version of this script, too, so fix it in both
places.

--BZR--
revision-id: timothy.smith@sun.com-20091103205028-afmvdk2r26eltv12
property-branch-nick: 50-bugfix
testament3-sha1: 2c8fee254a3456eb0a0da56bca4947652dc278bd