Bug #37183 insert ignore into .. select ... hangs after            deadlock was encountered
The bug is caused by an inconsistent handling of the IGNORE
clause.  A read from a const table caused a lock timeout
(ER_LOCK_TIMEOUT) in innodb.  Since the IGNORE clause was
given, the timeout was converted into a warning instead of
an error, thus not populating the diagnostics area.  When
innodb subsequently marked the transaction for rollback,
mysql asserted since the diag.area was empty.

This patch consists of only a test case, as the bug itself
was fixed by the patch for Bug #46539



--BZR--
revision-id: magne.mahre@sun.com-20091112114333-0ziiuhhc1m2pv7v4
property-branch-nick: mysql-5.1-bugteam-37183
testament3-sha1: 53c710904a35fc42b80bf61218df3673343946a4